requirements.txt
# flask
# flask-cors
# tensorflow-cpu==2.15.0
# numpy
# pillow
# langchain
# langchain-mistralai
# python-dotenv
# gunicorn


flask==3.1.3
flask-cors==6.0.2
tensorflow-cpu==2.15.0
numpy==1.26.4
pillow==12.2.0
langchain==1.2.15
langchain-core==1.2.31
langchain-mistralai==1.1.2
python-dotenv==1.2.2
gunicorn==25.3.0